a bit more documentation
