+++ src/components/ANOVA/ANOVA.js
@@ -1,11 +1,13 @@
 import { randomInt } from 'mathjs';
-import { useState, Fragment } from 'react';
+import { useState, Fragment, useEffect } from 'react';
+import { Button } from 'react-bootstrap';
 import FTest from './FTest';
 import PopulationRow from './PopulationRow';
 import PopulationSettings from './PopulationSettings';
+import SimulateType1Error from './SimulateType1Error';
 
 export default function ANOVA() {
-  const [showFTest, setShowFTest] = useState(false);
+  const [showResults, setShowResults] = useState(false);
   const [populations, setPopulations] = useState([
     {
       id: 1,
@@ -23,10 +25,16 @@ export default function ANOVA() {
     }
   ]);
 
+  useEffect(() => {
+    setShowResults(false)
+  }, [populations]);
+
+  const samples = populations.map(({ sample }) => sample.map(({ x }) => x));
+
   return (
     <>
-      <PopulationSettings populations={populations} setPopulations={setPopulations} setShowFTest={setShowFTest}/>
-      {showFTest && (
+      <PopulationSettings populations={populations} setPopulations={setPopulations}/>
+      {populations.every(({ data }) => data.length !== 0) && (
         <>
           {populations.map(({ id, data, sample, sampleSize }) => (
             <Fragment key={id}>
@@ -35,7 +43,21 @@ export default function ANOVA() {
             </Fragment>
           ))}
           <hr/>
-          <FTest populations={populations}/>
+          <Button
+            variant="outline-primary"
+            active={showResults}
+            onClick={() => setShowResults(true)}
+            disabled={samples.some((sample) => sample.length === 0)}
+          >
+            Run F-Test
+          </Button>
+          {showResults && (
+            <>
+              <FTest populations={populations} samples={samples}/>
+              <hr/>
+              <SimulateType1Error/>
+            </>
+          )}
         </>
       )}
     </>
